pyproject.toml
[project]
name = "facecrop"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.9,<3.12"
dependencies = [
    "insightface>=0.7.3",
    "numpy>=2.0.2",
    "onnxruntime>=1.20.1",
    "opencv-python>=4.12.0.88",
    "pillow>=11.3.0",
    "rich>=14.1.0",
    "tqdm>=4.67.1",
    "typer>=0.19.2",
]

[project.scripts]
facecrop = "main:main"
